requirements.txt
python-telegram-bot==20.8
motor==3.3.2
loguru==0.7.2
dnspython==2.6.0
pymongo==4.6.3
